fixup

+++ lib/fixup.rb
@@ -5,7 +5,7 @@
 FileUtils.rm_rf "#{HOME}/.cache/crewcache/manifest"
 
 # Remove install.sh provided path file since we supersede it.
-if File.exist?("#{CREW_PREFIX}/etc/env.d/00-path")
+if File.exist?("#{CREW_PREFIX}/etc/env.d/00-path") && File.exist?("#{CREW_PREFIX}/etc/env.d/path")
   puts "Removing #{CREW_PREFIX}/etc/env.d/path installed by the Chromebrew installer.\n".orange
   FileUtils.rm "#{CREW_PREFIX}/etc/env.d/path"
 end
